Even more refactoring - extracted a BaseCard class which is parent to skat.Card and doko.Card - extracted a BaseUI class which is parent to the client classes
- implemented message handling in SkatClient
- implemented sorting of cards in skat
