+++ Players.log
@@ -54,3 +54,21 @@
 2021-03-07 18:54:30(259899) - info - 26 - Players - Simon is Re. Reason: hat Kreuz Dame
 2021-03-07 18:57:13(162774) - info - 26 - Players - Volker is Re. Reason: hat Kreuz Dame
 2021-03-07 18:57:13(162774) - info - 26 - Players - Simon is Re. Reason: hat Kreuz Dame
+2021-03-27 16:34:50(209963) - info - 27 - Players - Volker is Re. Reason: hat Kreuz Dame
+2021-03-27 16:34:50(92847) - info - 27 - Players - Thomas is Re. Reason: hat Kreuz Dame
+2021-03-27 17:13:03(105724) - info - 27 - Players - Volker is Re. Reason: hat Kreuz Dame
+2021-03-27 17:13:03(76747) - info - 27 - Players - Simon is Re. Reason: hat Kreuz Dame
+2021-03-27 18:11:46(51088) - info - 27 - Players - Thomas is Re. Reason: hat Kreuz Dame
+2021-03-27 18:11:46(43518) - info - 27 - Players - Simon is Re. Reason: hat Kreuz Dame
+2021-03-27 18:17:44(51295) - info - 27 - Players - Volker is Re. Reason: hat Kreuz Dame
+2021-03-27 18:19:27(114967) - info - 27 - Players - Thomas is Re. Reason: hat Kreuz Dame
+2021-03-27 18:23:32(347543) - info - 27 - Players - Volker is Re. Reason: hat Kreuz Dame
+2021-03-27 18:23:32(391801) - info - 27 - Players - Robin is Re. Reason: hat Kreuz Dame
+2021-03-27 18:35:34(15902) - info - 27 - Players - Simon is Re. Reason: hat Kreuz Dame
+2021-03-27 18:35:34(14430) - info - 27 - Players - Robin is Re. Reason: hat Kreuz Dame
+2021-03-27 18:36:58(102972) - info - 27 - Players - Volker is Re. Reason: hat Kreuz Dame
+2021-03-27 18:36:58(84733) - info - 27 - Players - Robin is Re. Reason: hat Kreuz Dame
+2021-03-27 18:37:28(29284) - info - 27 - Players - Robin is Re. Reason: hat Kreuz Dame
+2021-03-27 18:40:58(341525) - info - 27 - Players - Thomas is Re. Reason: hat Kreuz Dame
+2021-03-27 18:40:58(210425) - info - 27 - Players - Robin is Re. Reason: hat Kreuz Dame
+2021-04-07 21:51:59(55296) - info - 28 - Players - Jens is Re. Reason: hat Kreuz Dame
