+++ src/main/java/BaseServer.java
@@ -1,0 +1,7 @@
+import base.Player;
+import base.messages.RequestObject;
+import base.messages.StartGame;
+
+public class BaseServer {
+
+}
