+++ src/main/java/IInputputHandler.java
@@ -1,3 +1,5 @@
+import base.messages.RequestObject;
+
 public interface IInputputHandler {
-    public void handleInput(String input);
+    public void handleInput(RequestObject input);
 }
